+++ cutils/src/main/java/com/carlos/cutils/constants/RegularConstants.kt
@@ -1,0 +1,16 @@
+package com.carlos.cutils.constants
+
+/**
+ * Created by Carlos on 2019-12-17.
+ */
+const val ONLY_ONE_CHINESE = "[\\u4e00-\\u9fa5]"
+const val ONLY_CHINESE = "[\\u4e00-\\u9fa5]+"
+
+const val ONLY_ONE_ENGLISH = "[a-zA-Z]"
+const val ONLY_ENGLISH = "[a-zA-Z]+"
+
+const val ONLY_ONE_NUMBER = "[0-9]"
+const val ONLY_NUMBER = "[0-9]+"
+
+const val ONE_CHINESE_OR_ENGLISH = "[a-zA-Z|\\u4e00-\\u9fa5]"
+const val CHINESE_ENGLISH = "[a-zA-Z|\\u4e00-\\u9fa5]+"