+++ cutils/src/main/java/com/carlos/cutils/extend/ActivityExtend.kt
@@ -1,0 +1,12 @@
+package com.carlos.cutils.extend
+
+import android.app.Activity
+import android.content.Intent
+
+/**
+ * Created by Carlos on 2019-12-17.
+ */
+fun <T : Activity> T.defaultStartActivity(clazz: Class<*>) {
+    val intent = Intent(this, clazz)
+    this.startActivity(intent)
+}