+++ cutils/src/main/java/com/carlos/cutils/view/CircleButton.kt
@@ -3,3 +3,6 @@ package com.carlos.cutils.view
 /**
  * Created by Carlos on 2019-05-17.
  */
+class CircleButton {
+
+}