+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* `19/12/25` Add RegularConstans class.
 * `19/12/16` Config jcenter and update Android Studio to 3.5.3. 
 * `19/12/08` Add EncodeUtils class. 
 * `19/12/06` Add CBaseEditActivity class.